fintools: Remove instances of hardcoded data format
Some places in the code relied on a specific order of data and metadata
fields. Here we decouple the data format from the implementation. To do
this we ended up generalizing the 'munge_csv_prepend_header' a bit and
renamed to 'munge_csv_metadata_header'.

Currently the metadata is just column and row count as well as minmax
data for specificed columns. The format is as follows:

<column count><sep><row count>
<arg #1 col min><sep><arg #1 col max>
<arg #2 col min><sep><arg #2 col max>
...

Note that metadata is separated from data by a newline. This is
documented somewhat more carefully in the comments.